build: squash into header-only
