+++ .github/workflows/lint.yml
@@ -24,4 +24,3 @@ jobs:
         run: |
           cd hypercomplex
           /home/runner/.local/bin/cpplint --quiet Hypercomplex.h
-          /home/runner/.local/bin/cpplint --quiet Hypercomplex.cpp
