+++ .github/workflows/unit.yml
@@ -43,7 +43,7 @@ jobs:
 
       - name: Compile Test Program
         working-directory: ${{env.working-directory}}
-        run: g++ -O0 -Wall --std=c++14 -o test test.cpp hypercomplex/Hypercomplex.cpp -lmpfr
+        run: g++ -O0 -Wall --std=c++14 -o test test.cpp -lmpfr
 
       - name: Execute Test Program
         working-directory: ${{env.working-directory}}
@@ -73,7 +73,7 @@ jobs:
 
       - name: Compile Test Program
         working-directory: ${{env.working-directory}}
-        run: g++ -O0 -Wall --std=c++14 --coverage -o test test.cpp hypercomplex/Hypercomplex.cpp -lmpfr
+        run: g++ -O0 -Wall --std=c++14 --coverage -o test test.cpp -lmpfr
 
       - name: Execute Test Program
         working-directory: ${{env.working-directory}}
@@ -115,7 +115,7 @@ jobs:
 
       - name: Compile Test Program
         working-directory: ${{env.working-directory}}
-        run: g++ -O0 -Wall --std=c++14 -o test test.cpp hypercomplex/Hypercomplex.cpp -lmpfr
+        run: g++ -O0 -Wall --std=c++14 -o test test.cpp -lmpfr
 
       - name: Analyze Test Program Execution
         working-directory: ${{env.working-directory}}
